RMQ_Consumers/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
lxml==5.3.0
multidict==6.1.0
odfpy==1.4.1
openai==0.28.0
pika==1.3.2
pillow==11.0.0
propcache==0.2.1
pydantic==2.10.4
pydantic_core==2.27.2
PyMySQL==1.1.1
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
XlsxWriter==3.2.0
yarl==1.18.3
